fix(finance): four small clean-ups the final review flagged
Cancelled money on the Money card. Every cancelled project's bar ended
in an unexplained grey gap, because received, outstanding and written
off do not add up to the contract when milestones were cancelled. The
project summary now carries cancelledPaise and the bar gets a
"Cancelled" slice. It is the part of cancelled milestones that was NEVER
collected, not their full amount: 7 of the 13 cancelled projects took
money before cancelling, that money is already in receivedPaise, and the
full amount overshot the contract on all 7. With the uncollected part,
received + outstanding + written off + cancelled equals the contract
exactly on all 13. Computed in the existing summary query — no migration.

No rupee amount on /finance/entries. Each row carried a rupee `amount`
beside `amountPaise`. Nothing read it, and two units of the same money
in one row is how a component ends up adding the wrong one.

Casts on three Receivables totals. totalOutstandingPaise, overduePaise
and noDueDatePaise now cast ::BIGINT like every other money aggregate.
The service already normalised them, so figures are unchanged.

"Advances paid" says what it covers. Its total spans every vendor, but
"Owing only" hides the vendors holding an advance; the tile now says so
while that filter is on, instead of showing a figure the list cannot
explain.

Verified: entries rows have amountPaise and no amount; bucket totals
unchanged as integers; a cancelled project that collected first reads
received 19,799,900 + cancelled 614,593 = contract 20,414,493; its Money
card bar reads 97% received and 3% cancelled, summing to 100, with a
"Cancelled ₹6,145.93" label; the Advances note switches with the filter.

+++ apps/backend/src/modules/finance/services/finance-ledger-queries.sql.ts
@@ -402,10 +402,10 @@ export const RECEIVABLES_BUCKETS_SQL = `
     COUNT(*) FILTER (WHERE v.days_overdue BETWEEN 61 AND 90)   AS "d61to90",
     COUNT(*) FILTER (WHERE v.days_overdue > 90)                AS "d90plus",
     COUNT(*)                                                   AS "all",
-    COALESCE(SUM(v.balance_paise), 0)                          AS "totalOutstandingPaise",
-    COALESCE(SUM(v.balance_paise) FILTER (WHERE v.days_overdue > 0), 0) AS "overduePaise",
+    COALESCE(SUM(v.balance_paise), 0)::BIGINT                  AS "totalOutstandingPaise",
+    COALESCE(SUM(v.balance_paise) FILTER (WHERE v.days_overdue > 0), 0)::BIGINT AS "overduePaise",
     COUNT(*) FILTER (WHERE v.due_date IS NULL)                          AS "noDueDate",
-    COALESCE(SUM(v.balance_paise) FILTER (WHERE v.due_date IS NULL), 0) AS "noDueDatePaise",
+    COALESCE(SUM(v.balance_paise) FILTER (WHERE v.due_date IS NULL), 0)::BIGINT AS "noDueDatePaise",
     COUNT(DISTINCT pr.id)                                              AS "recoveryProjects",
     -- Defect 5: a loan project with no lender milestone means the customer is
     -- being chased for the bank's share. Counted, never repaired — a 10/70/20
